--- app/(main)/components/Navbar.jsx
@@ -1,67 +1,0 @@
-'use client';
-
-import { useState } from 'react';
-import LoginPopup from './LoginPopup';
-
-export default function Navbar() {
-    const [showLogin, setShowLogin] = useState(false);
-
-    return (
-        <>
-            <nav className="navbar bg-base-100 shadow-sm sticky top-0 z-10">
-                <div className="navbar-start">
-                    <div className="dropdown">
-                        <div tabIndex={0} role="button" className="btn btn-ghost lg:hidden">
-                            <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M4 6h16M4 12h8m-8 6h16" />
-                            </svg>
-                        </div>
-                        <ul
-                            tabIndex={0}
-                            className="menu menu-sm dropdown-content bg-base-100 rounded-box z-1 mt-3 w-52 p-2 shadow"
-                        >
-                            <li><a>Item 1</a></li>
-                            <li>
-                                <a>Parent</a>
-                                <ul className="p-2">
-                                    <li><a>Submenu 1</a></li>
-                                    <li><a>Submenu 2</a></li>
-                                </ul>
-                            </li>
-                            <li><a>Item 3</a></li>
-                        </ul>
-                    </div>
-                    <a className="btn bg-transparent border-transparent shadow-none text-xl">Wivern</a>
-                </div>
-
-                <div className="navbar-center hidden lg:flex">
-                    <ul className="menu menu-horizontal px-1">
-                        <li><a>Home</a></li>
-                        <li><a>About us</a></li>
-                        <li><a>Become a Freelancer</a></li>
-                        <li><a></a></li>
-                        <li>
-                            <details>
-                                <summary>Contact</summary>
-                                <ul className="p-2">
-                                    <li><a>Instagram</a></li>
-                                    <li><a>Whatsapp</a></li>
-                                </ul>
-                            </details>
-                        </li>
-                    </ul>
-                </div>
-
-                <div className="navbar-end">
-                    <button
-                        className="btn bg-black text-white dark:bg-white dark:text-black"
-                        onClick={() => setShowLogin(true)}
-                    >
-                        Join
-                    </button>
-                </div>
-            </nav>
-            {showLogin && <LoginPopup onClose={() => setShowLogin(false)} />}
-        </>
-    );
-}
