+++ app/(main)/page.jsx
@@ -1,14 +1,7 @@
-"use client"
-
 import JoinSection from "./components/JoinSection";
 import EventsSection from "./components/EventSection";
-import MisiKami from "./components/Misi";
-import LoginPopup from "./components/LoginPopup";
-import React, { useState } from "react";
 
 export default function Home() {
-  const [showLogin, setShowLogin] = useState(false);
-
   return (
     <div>
 
@@ -26,7 +19,7 @@ export default function Home() {
             <h1 className="mb-5 text-5xl font-bold">
               Temukan Jasa Berkualitas Untuk Membantumu
             </h1>
-            <button className="btn shadow-none m-2" onClick={() => setShowLogin(true)}>Cari Jasa</button>
+            <button className="btn shadow-none m-2">Cari Jasa</button>
             <button className="btn shadow-none m-2">
               Daftar Sebagai Freelancer
             </button>
@@ -58,13 +51,91 @@ export default function Home() {
         </div>
       </div>
 
-      <MisiKami />
+      <div className="hero bg-gray-200 h-[91vh]">
+        <div className="p-8 max-h-[91vh]">
+          <div className="grid grid-cols-1 md:grid-cols-2 gap-8 max-w-6xl">
+            <div className="">
+              <div className="flex items-center mb-6">
+                <div className="w-0.5 h-4 bg-black mr-2"></div>
+                <h2 className="text-sm font-semibold uppercase">What We Do</h2>
+              </div>
+              <h1 className="text-3xl font-bold mb-4">Misi Kami</h1>
+              <p className="text-base text-gray-600 mb-8">
+                Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse
+                varius enim in eros elementum tristique.
+              </p>
+              <div className="mb-4">
+                <div className="flex items-center mb-2">
+                  <img
+                    src="https://via.placeholder.com/20x20"
+                    alt="Family Icon"
+                    className="w-5 h-5 mr-2"
+                  />
+                  <h3 className="text-lg font-bold">Family support</h3>
+                </div>
+                <p className="text-base text-gray-600">
+                  Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse
+                  varius enim in eros.
+                </p>
+              </div>
+              <div className="mb-4">
+                <div className="flex items-center mb-2">
+                  <img
+                    src="https://via.placeholder.com/20x20"
+                    alt="Health Icon"
+                    className="w-5 h-5 mr-2"
+                  />
+                  <h3 className="text-lg font-bold">Health benefits</h3>
+                </div>
+                <p className="text-base text-gray-600">
+                  Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse
+                  varius enim in eros.
+                </p>
+              </div>
+              <div className="mb-4">
+                <div className="flex items-center mb-2">
+                  <img
+                    src="https://via.placeholder.com/20x20"
+                    alt="Scholarship Icon"
+                    className="w-5 h-5 mr-2"
+                  />
+                  <h3 className="text-lg font-bold">Scholarships</h3>
+                </div>
+                <p className="text-base text-gray-600">
+                  Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse
+                  varius enim in eros.
+                </p>
+              </div>
+              <div>
+                <div className="flex items-center mb-2">
+                  <img
+                    src="https://via.placeholder.com/20x20"
+                    alt="Therapy Icon"
+                    className="w-5 h-5 mr-2"
+                  />
+                  <h3 className="text-lg font-bold">Therapy</h3>
+                </div>
+                <p className="text-base text-gray-600">
+                  Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse
+                  varius enim in eros.
+                </p>
+              </div>
+            </div>
+            <div className="flex justify-center items-center">
+              <img
+                src="/home2.png"
+                alt="Person Working on Laptop"
+                className="rounded-md max-w-md max-h-[80vh]"
+              />
+            </div>
+          </div>
+        </div>
+      </div>
       
       <div>
         <JoinSection />
         <EventsSection />
       </div>
-      {showLogin && <LoginPopup onClose={() => setShowLogin(false)}/>}
 
     </div>
